requirements.txt
thrift==0.9.3
requests==2.9.1
kafka-python==0.9.5
